Fixed bug (issue 2), fixed memory leak
- Fixed bug where commands would get previous commands args
I believe this was due to the fact that exec* was trying to
accsess recently free'd memory when given the whole value of argv.
This was fixed by passing the argument count to the call function.
This patch fixes bug 2: https://github.com/mckenney5/qsh/issues/2

- Fixed memory leak
The for loop that was used to free memory was missing the last mem block
due to using i < argc  instead of  i <= argc

- Fixed commands running twice
The program was running commands and/or allocating memory twice.
This was due to the fact that fork() was called too early.
To fix, a function named call was made that handled the actual
exec* work, instead of having all in run.